requirements.txt
# Core Dependencies
streamlit==1.31.0
chromadb>=1.5.2,<2.0.0
sentence-transformers==2.3.1
ollama>=0.6.0,<1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
PyPDF2==3.0.1
pdfplumber==0.10.3

# AWS Integration
boto3==1.34.34
awscli==1.32.34

# System Monitoring
psutil==5.9.8

# Utilities
python-dotenv==1.0.1
requests>=2.32.5,<3.0.0

# API Backend (React bridge)
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.0.0,<3.0.0

# Auth & Database
sqlalchemy>=2.0.0,<3.0.0
bcrypt>=4.0.0,<5.0.0
PyJWT>=2.8.0,<3.0.0

# LangChain / AI modules
langchain>=1.2.10
langchain-chroma==1.1.0
langchain-huggingface==1.2.0
langchain-ollama==1.0.1
langchain-core==1.2.16
langchain-text-splitters==1.1.1
langchain-community==0.4.1
